# api.yaml
openapi: 3.0.0
info:
  title: Vehicle Parking App Lot Management API
  version: 1.0.0
  description: API for user management and parking lot operations

servers:
  - url: http://localhost:5000/api

paths:
  /login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login success
        '400':
          description: Incorrect password
        '404':
          description: User not found
        '500':
          description: Internal server error

  /register:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                fullname:
                  type: string
                dob:
                  type: string
      responses:
        '201':
          description: Registration successful
        '400':
          description: Username already exists
        '500':
          description: Internal server error

  /update_user/:
    put:
      summary: Update user details
      parameters:
        - in: query
          name: user_id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
                dob:
                  type: string
      responses:
        '200':
          description: User updated successfully
        '400':
          description: Incorrect user id
        '500':
          description: Internal server error

  /delete_user/:
    delete:
      summary: Delete user
      parameters:
        - in: query
          name: user_id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: User deleted successfully
        '400':
          description: Incorrect user id
        '500':
          description: Internal server error

  /parkinglots:
    get:
      summary: List all parking lots
      responses:
        '200':
          description: Returns a list of parking lots
        '500':
          description: Internal server error
    post:
      summary: Create a new parking lot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prime_location_name:
                  type: string
                address:
                  type: string
                pin_code:
                  type: string
                price:
                  type: number
                max_no_spots:
                  type: integer
      responses:
        '201':
          description: Parking lot created
        '500':
          description: Internal server error

  /parkinglots/{lot_id}:
    get:
      summary: Get parking lot details
      parameters:
        - in: path
          name: lot_id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Parking lot details returned
        '404':
          description: Not found
        '500':
          description: Internal server error
    put:
      summary: Update parking lot
      parameters:
        - in: path
          name: lot_id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prime_location_name:
                  type: string
                address:
                  type: string
                pin_code:
                  type: string
                price:
                  type: number
                max_no_spots:
                  type: integer
      responses:
        '200':
          description: Lot updated
        '500':
          description: Internal server error
    delete:
      summary: Delete parking lot
      parameters:
        - in: path
          name: lot_id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Lot deleted
        '400':
          description: Cannot delete lot with occupied spots
        '500':
          description: Internal server error

  /parkinglots/{lot_id}/spots:
    get:
      summary: List spots for parking lot
      parameters:
        - in: path
          name: lot_id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: List of spots
        '500':
          description: Internal server error

  /reserve:
    post:
      summary: Reserve a spot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                lot_id:
                  type: integer
                vehicle_no:
                  type: string
      responses:
        '201':
          description: Spot booked
        '400':
          description: No spots available
        '500':
          description: Internal server error

  /release:
    post:
      summary: Release a spot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reservation_id:
                  type: integer
      responses:
        '200':
          description: Spot released with total cost
        '400':
          description: Spot already released
        '500':
          description: Internal server error

  /user/{user_id}/reservations:
    get:
      summary: Get reservations for a user
      parameters:
        - in: path
          name: user_id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: List of reservations
        '500':
          description: Internal server error